+++ domain/src/main/kotlin/org/monsing/BaseEntity.kt
@@ -1,6 +1,11 @@
 package org.monsing
 
-import jakarta.persistence.*
+
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.MappedSuperclass
 import org.springframework.data.annotation.CreatedDate
 import org.springframework.data.annotation.LastModifiedDate
 import java.time.LocalDateTime
