+++ domain/src/main/kotlin/org/monsing/course/ClassNumber.kt
@@ -1,0 +1,11 @@
+
+
+package org.monsing.course
+
+import jakarta.persistence.Embeddable
+
+@Embeddable
+class ClassNumber(
+    val value: Int
+) {
+}
