+++ domain/src/main/kotlin/org/monsing/course/CourseScheduleRepository.kt
@@ -1,0 +1,8 @@
+package org.monsing.course
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface CourseScheduleRepository : JpaRepository<CourseSchedule, Long> {
+}
