+++ domain/src/main/kotlin/org/monsing/course/DayOfWeek.kt
@@ -1,0 +1,11 @@
+package org.monsing.course
+
+enum class DayOfWeek {
+    MONDAY,
+    TUESDAY,
+    WEDNESDAY,
+    THURSDAY,
+    FRIDAY,
+    SATURDAY,
+    SUNDAY
+}
