+++ domain/src/main/kotlin/org/monsing/course/Price.kt
@@ -1,0 +1,11 @@
+package org.monsing.course
+
+import jakarta.persistence.Column
+import jakarta.persistence.Embeddable
+
+@Embeddable
+class Price(
+    @Column(name = "price")
+    val value: Int
+) {
+}
