+++ domain/src/main/kotlin/org/monsing/member/FeedbackRepository.kt
@@ -1,0 +1,8 @@
+package org.monsing.member
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface FeedbackRepository : JpaRepository<Feedback, Long> {
+}
