+++ domain/src/main/kotlin/org/monsing/member/FeedbackStatus.kt
@@ -1,0 +1,6 @@
+package org.monsing.member
+
+enum class FeedbackStatus {
+    REQUESTED,
+    COMPLETED
+}
