+++ domain/src/main/kotlin/org/monsing/member/GenderType.kt
@@ -1,0 +1,7 @@
+package org.monsing.member
+
+enum class GenderType {
+    MALE,
+    FEMALE,
+    OTHER
+}
