+++ domain/src/main/kotlin/org/monsing/member/MemberType.kt
@@ -1,0 +1,6 @@
+package org.monsing.member
+
+enum class MemberType {
+    STUDENT,
+    TEACHER
+}
