+++ domain/src/main/kotlin/org/monsing/member/Nickname.kt
@@ -1,0 +1,12 @@
+package org.monsing.member
+
+import jakarta.persistence.Column
+import jakarta.persistence.Embeddable
+
+@Embeddable
+class Nickname(
+    @Column(name = "nickname")
+    val value: String
+) {
+
+}
