+++ domain/src/main/kotlin/org/monsing/member/RatingScore.kt
@@ -1,0 +1,12 @@
+package org.monsing.member
+
+import jakarta.persistence.Column
+import jakarta.persistence.Embeddable
+
+@Embeddable
+class RatingScore(
+    @Column(name = "score")
+    val value: Int
+) {
+
+}
