+++ domain/src/main/kotlin/org/monsing/member/RatingType.kt
@@ -1,0 +1,7 @@
+package org.monsing.member
+
+enum class RatingType {
+    KINDNESS,
+    KNOWLEDGE,
+    PASSION
+}
