+++ domain/src/main/kotlin/org/monsing/member/Record.kt
@@ -1,0 +1,15 @@
+package org.monsing.member
+
+import jakarta.persistence.Entity
+import jakarta.persistence.ManyToOne
+import org.monsing.BaseEntity
+import java.net.URL
+
+@Entity
+class Record(
+    @ManyToOne
+    val student: Student,
+
+    val url: URL
+) : BaseEntity() {
+}
