+++ domain/src/main/kotlin/org/monsing/member/StrongSideType.kt
@@ -1,0 +1,7 @@
+package org.monsing.member
+
+enum class StrongSideType {
+    PITCH,
+    FIX,
+    RHYTHM
+}
