+++ domain/src/main/kotlin/org/monsing/member/StudentRepository.kt
@@ -1,0 +1,8 @@
+package org.monsing.member
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface StudentRepository : JpaRepository<Student, Long> {
+}
