+++ domain/src/main/kotlin/org/monsing/report/ReportImageRepository.kt
@@ -1,0 +1,8 @@
+package org.monsing.report
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface ReportImageRepository : JpaRepository<ReportImage, Long> {
+}
