+++ domain/src/main/kotlin/org/monsing/report/ReportType.kt
@@ -1,0 +1,7 @@
+package org.monsing.report
+
+enum class ReportType {
+    INSULT,
+    SPAM,
+    SEXUAL
+}
